Team 3 Updates

+++ src/main/java/com/nucleus/loanapplications/model/LoanApplications.java
@@ -37,7 +37,7 @@ public class LoanApplications {
     private Integer tenure;
 
     @Column(name = "rate",nullable = false)
-    private Integer rate;
+    private Double rate;
 
     @Column(name = "agreement_date",nullable = false)
     private LocalDate agreementDate;
@@ -94,11 +94,11 @@ public class LoanApplications {
         this.tenure = tenure;
     }
 
-    public Integer getRate() {
+    public Double getRate() {
         return rate;
     }
 
-    public void setRate(Integer rate) {
+    public void setRate(Double rate) {
         this.rate = rate;
     }
